+++ patches/tModLoader/Terraria/ModLoader/ItemLoader.cs
@@ -555,13 +555,13 @@ namespace Terraria.ModLoader
 		}
 
 
-		private delegate void DelegateModifyWeaponCrit(Item item, Player player, ref int crit);
+		private delegate void DelegateModifyWeaponCrit(Item item, Player player, ref float crit);
 		private static HookList HookModifyWeaponCrit = AddHook<DelegateModifyWeaponCrit>(g => g.ModifyWeaponCrit);
 
 		/// <summary>
 		/// Calls ModItem.ModifyWeaponCrit, then all GlobalItem.ModifyWeaponCrit hooks.
 		/// </summary>
-		public static void ModifyWeaponCrit(Item item, Player player, ref int crit) {
+		public static void ModifyWeaponCrit(Item item, Player player, ref float crit) {
 			if (item.IsAir)
 				return;
